# docs/v3/openapi.yaml
openapi: 3.0.1
info:
  contact: {}
  description: REST API for managing software licenses
  title: License Server API
  version: 0.1.0
servers:
- url: http://localhost:8080/api/v1
paths:
  /admin/products:
    get:
      description: retrieves all products
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-array_model_Product'
          description: OK
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      security:
      - ApiKey: []
      summary: Get all products
      tags:
      - products
    post:
      description: creates a new product
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/model.Product'
        description: new product info
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-model_Product'
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      security:
      - ApiKey: []
      summary: Add a product
      tags:
      - products
      x-codegen-request-body-name: data
  /admin/products/:id:
    delete:
      description: deletes a specific product
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-model_Product'
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      security:
      - ApiKey: []
      summary: Delete a product
      tags:
      - products
    get:
      description: retrieve a specific product
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-model_Product'
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      security:
      - ApiKey: []
      summary: Get a product
      tags:
      - products
    put:
      description: updates a specific product
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/model.Product'
        description: updated product info
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-model_Product'
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      security:
      - ApiKey: []
      summary: Update a product
      tags:
      - products
      x-codegen-request-body-name: data
...
  /auth/register:
    post:
      description: registers new application user
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/controller.AuthRequest'
        description: new user info
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-model_User'
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      summary: Register a user
      tags:
      - auth
      x-codegen-request-body-name: data
  /auth/token:
    post:
      description: creates a new API key
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/controller.AuthRequest'
        description: returning user info
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-model_DisplayAPIKey'
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      summary: Create a token
      tags:
      - auth
      x-codegen-request-body-name: data
  /licenses/issue:
    post:
      description: issues a new product license
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/model.LicenseRequest'
        description: new license info
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-any'
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      security:
      - ApiKey: []
      summary: Issue a license
      tags:
      - licenses
      x-codegen-request-body-name: data
  /licenses/validate:
    post:
      description: validates a product license
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPResponse-any'
          description: OK
        "400":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Bad Request
        "500":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/httputil.HTTPError'
          description: Internal Server Error
      security:
      - ApiKey: []
      summary: Validate a license
      tags:
      - licenses
components:
  schemas:
    controller.AuthRequest:
      properties:
        email:
          type: string
        name:
          type: string
        password:
          type: string
      required:
      - email
      type: object
...
    httputil.HTTPError:
      properties:
        code:
          example: 400
          type: integer
        message:
          example: status bad request
          type: string
      type: object
    httputil.HTTPResponse-any:
      example:
        data: "{}"
        message: message
      properties:
        data:
          type: object
        message:
          type: string
      type: object
    httputil.HTTPResponse-array_model_Product:
      example:
        data:
        - features:
          - updated_at: 5
            product_id: product_id
            name: name
            created_at: 1
            id: id
            deleted_at: 5
          - updated_at: 5
            product_id: product_id
            name: name
            created_at: 1
            id: id
            deleted_at: 5
          licenses:
          - updated_at: 9
            product_id: product_id
            name: name
            created_at: 2
            id: id
            deleted_at: 7
          - updated_at: 9
            product_id: product_id
            name: name
            created_at: 2
            id: id
            deleted_at: 7
          updated_at: 3
          name: name
          created_at: 0
          id: id
          deleted_at: 6
          version: version
        - features:
          - updated_at: 5
            product_id: product_id
            name: name
            created_at: 1
            id: id
            deleted_at: 5
          - updated_at: 5
            product_id: product_id
            name: name
            created_at: 1
            id: id
            deleted_at: 5
          licenses:
          - updated_at: 9
            product_id: product_id
            name: name
            created_at: 2
            id: id
            deleted_at: 7
          - updated_at: 9
            product_id: product_id
            name: name
            created_at: 2
            id: id
            deleted_at: 7
          updated_at: 3
          name: name
          created_at: 0
          id: id
          deleted_at: 6
          version: version
        message: message
      properties:
        data:
          items:
            $ref: '#/components/schemas/model.Product'
          type: array
        message:
          type: string
      type: object
...
    httputil.HTTPResponse-model_DisplayAPIKey:
      example:
        data:
          authentication_scopes: authentication_scopes
          expires_at: expires_at
          updated_at: 1
          user_id: user_id
          created_at: 0
          id: id
          deleted_at: 6
          key: key
        message: message
      properties:
        data:
          $ref: '#/components/schemas/model.DisplayAPIKey'
        message:
          type: string
      type: object
    httputil.HTTPResponse-model_Product:
      example:
        data:
          features:
          - updated_at: 5
            product_id: product_id
            name: name
            created_at: 1
            id: id
            deleted_at: 5
          - updated_at: 5
            product_id: product_id
            name: name
            created_at: 1
            id: id
            deleted_at: 5
          licenses:
          - updated_at: 9
            product_id: product_id
            name: name
            created_at: 2
            id: id
            deleted_at: 7
          - updated_at: 9
            product_id: product_id
            name: name
            created_at: 2
            id: id
            deleted_at: 7
          updated_at: 3
          name: name
          created_at: 0
          id: id
          deleted_at: 6
          version: version
        message: message
      properties:
        data:
          $ref: '#/components/schemas/model.Product'
        message:
          type: string
      type: object
    httputil.HTTPResponse-model_User:
      example:
        data:
          updated_at: 1
          name: name
          created_at: 0
          id: id
          deleted_at: 6
          email: email
        message: message
      properties:
        data:
          $ref: '#/components/schemas/model.User'
        message:
          type: string
      type: object
...
    model.DisplayAPIKey:
      example:
        authentication_scopes: authentication_scopes
        expires_at: expires_at
        updated_at: 1
        user_id: user_id
        created_at: 0
        id: id
        deleted_at: 6
        key: key
      properties:
        authentication_scopes:
          type: string
        created_at:
          type: integer
        deleted_at:
          type: integer
        expires_at:
          type: string
        id:
          type: string
        key:
          type: string
        updated_at:
          type: integer
        user_id:
          type: string
      type: object
    model.License:
      example:
        updated_at: 9
        product_id: product_id
        name: name
        created_at: 2
        id: id
        deleted_at: 7
      properties:
        created_at:
          type: integer
        deleted_at:
          type: integer
        id:
          type: string
        name:
          type: string
        product_id:
          type: string
        updated_at:
          type: integer
      type: object
    model.LicenseRequest:
      properties:
        features:
          additionalProperties:
            type: boolean
          type: object
        productId:
          type: string
        startDate:
          type: integer
        userId:
          type: string
      required:
      - productId
      type: object
    model.Product:
      example:
        features:
        - updated_at: 5
          product_id: product_id
          name: name
          created_at: 1
          id: id
          deleted_at: 5
        - updated_at: 5
          product_id: product_id
          name: name
          created_at: 1
          id: id
          deleted_at: 5
        licenses:
        - updated_at: 9
          product_id: product_id
          name: name
          created_at: 2
          id: id
          deleted_at: 7
        - updated_at: 9
          product_id: product_id
          name: name
          created_at: 2
          id: id
          deleted_at: 7
        updated_at: 3
        name: name
        created_at: 0
        id: id
        deleted_at: 6
        version: version
      properties:
        created_at:
          type: integer
        deleted_at:
          type: integer
        features:
          items:
            $ref: '#/components/schemas/model.ProductFeature'
          type: array
        id:
          type: string
        licenses:
          items:
            $ref: '#/components/schemas/model.License'
          type: array
        name:
          type: string
        updated_at:
          type: integer
        version:
          type: string
      type: object
    model.ProductFeature:
      example:
        updated_at: 5
        product_id: product_id
        name: name
        created_at: 1
        id: id
        deleted_at: 5
      properties:
        created_at:
          type: integer
        deleted_at:
          type: integer
        id:
          type: string
        name:
          type: string
        product_id:
          type: string
        updated_at:
          type: integer
      type: object
    model.User:
      example:
        updated_at: 1
        name: name
        created_at: 0
        id: id
        deleted_at: 6
        email: email
      properties:
        created_at:
          type: integer
        deleted_at:
          type: integer
        email:
          type: string
        id:
          type: string
        name:
          type: string
        updated_at:
          type: integer
      type: object
...
  securitySchemes:
    ApiKey:
      in: header
      name: X-API-KEY
      type: apiKey
...